deepfakeforsound/requirements_fixed.txt
# Core ML/DL libraries
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.8.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0

# Data handling
pandas>=1.4.0
scikit-learn>=1.1.0
matplotlib>=3.5.0

# Utilities
tqdm>=4.60.0
tensorboard>=2.8.0
pyyaml>=6.0

# Export and deployment
onnx>=1.12.0
onnxruntime>=1.12.0
fastapi>=0.85.0
uvicorn>=0.18.0
python-multipart>=0.0.5
